feat: 클래스 분리
